+++ dask/array/creation.py
@@ -10,7 +10,7 @@ from tlz import sliding_window
 
 from ..base import tokenize
 from ..highlevelgraph import HighLevelGraph
-from ..utils import derived_from
+from ..utils import cached_cumsum, derived_from
 from . import chunk
 from .core import (
     Array,
@@ -19,7 +19,6 @@ from .core import (
     blockwise,
     broadcast_arrays,
     broadcast_to,
-    cached_cumsum,
     concatenate,
     normalize_chunks,
     stack,
